refactor(analytics): split posthog events into focused modules
Decompose the 685-LOC events.ts into a slim public-facade plus five
focused sibling modules. Public API preserved via re-exports; no
behavior change.

- eventsCore.ts (172 LOC): trackLayoutSnapshot, track3DPreview,
  trackLayoutAction, trackFillOperation, trackPaintMode,
  trackBinCreated + MILESTONE_THRESHOLDS + checkEngagementMilestones.
- eventsHeartbeat.ts (120 LOC): activity context + heartbeat payload
  builder + trackHeartbeat.
- eventsPerson.ts (136 LOC): computeEngagementTier,
  updatePersonProperties, markFeatureUsed.
- eventsErrors.ts (105 LOC): captureException + 4 trackXFailure
  helpers (save, share, 3D render, template).
- eventsPerformance.ts (92 LOC): WASM threading, cache stats, kernel
  timing, baseplate preview timing.
- events.ts (149 LOC): facade re-exports + the standalone PWA install,
  UTM capture, and gallery open/close trackers.
